requirements.txt
pyserial~=3.5
requests~=2.26.0
